* sysdeps/unix/sysv/linux/aarch64/dl-auxv.h 	(__libc_arm64_midr): Change to int. 	* sysdeps/unix/sysv/linux/aarch64/dl-sysdep.c 	(__libc_arm64_midr): Likewise.

+++ sysdeps/unix/sysv/linux/aarch64/dl-auxv.h
@@ -17,7 +17,7 @@
    <http://www.gnu.org/licenses/>.  */
 
 
-extern long long __libc_arm64_midr;
+extern int __libc_arm64_midr;
 
 #define DL_PLATFORM_AUXV				\
       case AT_ARM64_MIDR:				\
